refactor: migrate vcs store to pinia (#1065)
* refactor: migrate vcs store to pinia

* fix: fixed the path

* refactor: rename getVcs to getVCS

* refactor: rename useVcsStore to useVCSStore
