+++ frontend/src/components/RepositoryVCSProviderPanel.vue
@@ -38,12 +38,11 @@ export default { name: "RepositoryVCSProviderPanel" };
 </script>
 
 <script setup lang="ts">
-import { useStore } from "vuex";
 import { reactive, computed, watchEffect, onUnmounted, onMounted } from "vue";
 import isEmpty from "lodash-es/isEmpty";
 import { OAuthWindowEventPayload, openWindowForOAuth, VCS } from "../types";
 import { isOwner } from "../utils";
-import { pushNotification, useCurrentUser } from "@/store";
+import { pushNotification, useCurrentUser, useVCSStore } from "@/store";
 
 interface LocalState {
   selectedVCS?: VCS;
@@ -55,13 +54,13 @@ const emit = defineEmits<{
   (event: "set-code", payload: string): void;
 }>();
 
-const store = useStore();
+const vcsStore = useVCSStore();
 const state = reactive<LocalState>({});
 
 const currentUser = useCurrentUser();
 
 const prepareVCSList = () => {
-  store.dispatch("vcs/fetchVCSList");
+  vcsStore.fetchVCSList();
 };
 
 watchEffect(prepareVCSList);
@@ -74,7 +73,7 @@ onUnmounted(() => {
 });
 
 const vcsList = computed(() => {
-  return store.getters["vcs/vcsList"]();
+  return vcsStore.getVCSList();
 });
 
 const eventListener = (event: Event) => {
