+++ frontend/src/components/VCSSetupWizard.vue
@@ -34,7 +34,6 @@ import {
   defineComponent,
 } from "vue";
 import { useRouter } from "vue-router";
-import { useStore } from "vuex";
 import isEmpty from "lodash-es/isEmpty";
 import { BBStepTabItem } from "../bbkit/types";
 import VCSProviderBasicInfoPanel from "./VCSProviderBasicInfoPanel.vue";
@@ -51,7 +50,7 @@ import {
 } from "../types";
 import { isUrl } from "../utils";
 import { useI18n } from "vue-i18n";
-import { pushNotification, useOAuthStore } from "@/store";
+import { pushNotification, useOAuthStore, useVCSStore } from "@/store";
 
 const BASIC_INFO_STEP = 0;
 const OAUTH_INFO_STEP = 1;
@@ -72,8 +71,8 @@ export default defineComponent({
   },
   setup() {
     const { t } = useI18n();
-    const store = useStore();
     const router = useRouter();
+    const vcsStore = useVCSStore();
 
     const stepList: BBStepTabItem[] = [
       { title: t("version-control.setting.add-git-provider.basic-info.self") },
@@ -205,7 +204,7 @@ export default defineComponent({
       const vcsCreate: VCSCreate = {
         ...state.config,
       };
-      store.dispatch("vcs/createVCS", vcsCreate).then((vcs: VCS) => {
+      vcsStore.createVCS(vcsCreate).then((vcs: VCS) => {
         allowChangeCallback();
         router.push({
           name: "setting.workspace.version-control",
